Add route to remove movies and episodes
