+++ app/internal/server/routes.go
@@ -19,20 +19,22 @@ func (s *Server) httpServer(log *logrus.Entry) *http.Server {
 
 	mux := mux.NewRouter()
 
-	mux.HandleFunc("/movies", s.movieIds).Name("MoviesListIDs")
-	mux.HandleFunc("/movies/{id}", s.getMovieDetails).Name("GetMovieDetails")
+	mux.HandleFunc("/movies", s.movieIds).Name("MoviesListIDs").Methods("GET")
+	mux.HandleFunc("/movies/{id}", s.getMovieDetails).Name("GetMovieDetails").Methods("GET")
+	mux.HandleFunc("/movies/{id}", s.deleteMovie).Name("DeleteMovie").Methods("DELETE")
 
-	mux.HandleFunc("/shows", s.showIds).Name("ShowsListIDs")
-	mux.HandleFunc("/shows/{id}", s.getShowDetails).Name("GetShowDetails")
-	mux.HandleFunc("/shows/{id}/{season:[0-9]+}/{episode:[0-9]+}", s.getShowEpisodeIDDetails).Name("GetShowEpisodeIDDetails")
-	mux.HandleFunc("/wishlist", s.wishlist).Name("Wishlist")
+	mux.HandleFunc("/shows", s.showIds).Name("ShowsListIDs").Methods("GET")
+	mux.HandleFunc("/shows/{id}", s.getShowDetails).Name("GetShowDetails").Methods("GET")
+	mux.HandleFunc("/shows/{id}/{season:[0-9]+}/{episode:[0-9]+}", s.getShowEpisodeIDDetails).Name("GetShowEpisodeIDDetails").Methods("GET")
+	mux.HandleFunc("/shows/{id}/{season:[0-9]+}/{episode:[0-9]+}", s.deleteEpisode).Name("DeleteEpisode").Methods("DELETE")
+	mux.HandleFunc("/wishlist", s.wishlist).Name("Wishlist").Methods("GET")
 
-	mux.HandleFunc("/torrents", s.addTorrent).Methods("POST").Name("TorrentsAdd")
+	mux.HandleFunc("/torrents", s.addTorrent).Name("TorrentsAdd").Methods("POST")
 
 	if s.config.HTTPServer.ServeFiles {
 		log.Debug("server will be serving files")
-		mux.HandleFunc("/shows/{id}/{season:[0-9]+}/{episode:[0-9]+}/download", s.serveShow).Name("ServeShowsByIDs")
-		mux.HandleFunc("/movies/{id}/download", s.serveMovie).Name("ServeMoviesByIDs")
+		mux.HandleFunc("/shows/{id}/{season:[0-9]+}/{episode:[0-9]+}/download", s.serveShow).Name("ServeShowsByIDs").Methods("GET")
+		mux.HandleFunc("/movies/{id}/download", s.serveMovie).Name("ServeMoviesByIDs").Methods("GET")
 	}
 
 	n := negroni.New()
